Add Quartz
- Change DateTime on Unix format
- Add scheduler micrisercive based on Quartz
- Move all enums to Core
- Save search settings from tg to searchSettings db

+++ FlyAnytime.Telegram/Bot/Conversations/UpdateSettingsConversation/Steps/UpdatePriceAndCountryStep.cs
@@ -66,7 +66,7 @@ namespace FlyAnytime.Telegram.Bot.Conversations.UpdateSettingsConversation.Steps
 
             var message = new AddOrUpdateBaseSearchSettingsMessage(
                 ChatId, 
-                double.Parse(price), 
+                decimal.Parse(price), 
                 chat.SearchSettings.ChatCountry.Code,
                 chat.SearchSettings.ChatCity.Code,
                 countryToCode);
